+++ src/games/registry.ts
@@ -7,6 +7,7 @@
 
 import { IGame } from '../core/types';
 import { CallStackGame } from './call-stack/config';
+import { CiBirdGame } from './ci-bird/config';
 import { DebugSnakeGame } from './debug-snake/config';
 import { KernelPanicGame } from './kernel-panic/config';
 import { MergeConflictGame } from './merge-conflict/config';
@@ -19,6 +20,7 @@ import { WhackABugGame } from './whack-a-bug/config';
 export const AllGames: IGame[] = [
     DebugSnakeGame,
     WhackABugGame,
+    CiBirdGame,
     CallStackGame,
     MergeConflictGame,
     KernelPanicGame